requirements.txt
pandas
SQLAlchemy
PyMySQL
requests
python-dotenv
tqdm